requirement.txt
autopep8==1.5
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
PyJWT==1.7.1
python-dotenv==0.12.0
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
